GrantMigratorRoleCommandBaseHandler
