+++ src/Octoshift/Commands/GrantMigratorRoleCommandBase.cs
@@ -8,15 +8,13 @@ namespace OctoshiftCLI.Commands;
 
 public class GrantMigratorRoleCommandBase : Command
 {
-    private readonly OctoLogger _log;
-    private readonly ITargetGithubApiFactory _githubApiFactory;
+    protected GrantMigratorRoleCommandBaseHandler BaseHandler { get; init; }
 
     public GrantMigratorRoleCommandBase(OctoLogger log, ITargetGithubApiFactory githubApiFactory) : base(
         name: "grant-migrator-role",
         description: "Allows an organization admin to grant a USER or TEAM the migrator role for a single GitHub organization. The migrator role allows the role assignee to perform migrations into the target organization.")
     {
-        _log = log;
-        _githubApiFactory = githubApiFactory;
+        BaseHandler = new GrantMigratorRoleCommandBaseHandler(log, githubApiFactory);
     }
 
     protected virtual Option<string> GithubOrg { get; } = new("--github-org") { IsRequired = true };
@@ -36,51 +34,6 @@ public class GrantMigratorRoleCommandBase : Command
         AddOption(ActorType);
         AddOption(GithubPat);
         AddOption(Verbose);
-    }
-
-    public async Task Handle(GrantMigratorRoleCommandArgs args)
-    {
-        if (args is null)
-        {
-            throw new ArgumentNullException(nameof(args));
-        }
-
-        _log.Verbose = args.Verbose;
-
-        _log.LogInformation("Granting migrator role ...");
-        _log.LogInformation($"{GithubOrg.GetLogFriendlyName()}: {args.GithubOrg}");
-        _log.LogInformation($"{Actor.GetLogFriendlyName()}: {args.Actor}");
-
-        args.ActorType = args.ActorType?.ToUpper();
-        _log.LogInformation($"{ActorType.GetLogFriendlyName()}: {args.ActorType}");
-
-        if (args.ActorType is "TEAM" or "USER")
-        {
-            _log.LogInformation("Actor type is valid...");
-        }
-        else
-        {
-            _log.LogError("Actor type must be either TEAM or USER.");
-            return;
-        }
-
-        if (args.GithubPat is not null)
-        {
-            _log.LogInformation($"{GithubPat.GetLogFriendlyName()}: ***");
-        }
-
-        var githubApi = _githubApiFactory.Create(targetPersonalAccessToken: args.GithubPat);
-        var githubOrgId = await githubApi.GetOrganizationId(args.GithubOrg);
-        var success = await githubApi.GrantMigratorRole(githubOrgId, args.Actor, args.ActorType);
-
-        if (success)
-        {
-            _log.LogSuccess($"Migrator role successfully set for the {args.ActorType} \"{args.Actor}\"");
-        }
-        else
-        {
-            _log.LogError($"Migrator role couldn't be set for the {args.ActorType} \"{args.Actor}\"");
-        }
     }
 }
