+++ src/OctoshiftCLI.Tests/Octoshift/Commands/GrantMigratorRoleCommandBaseTests.cs
@@ -13,7 +13,7 @@ public class GrantMigratorRoleCommandBaseTests
     private readonly Mock<ITargetGithubApiFactory> _mockGithubApiFactory = new();
     private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
 
-    private readonly GrantMigratorRoleCommandBase _command;
+    private readonly GrantMigratorRoleCommandBaseHandler _command;
 
     private const string GITHUB_ORG = "FooOrg";
     private const string ACTOR = "foo-actor";
@@ -21,7 +21,7 @@ public class GrantMigratorRoleCommandBaseTests
 
     public GrantMigratorRoleCommandBaseTests()
     {
-        _command = new GrantMigratorRoleCommandBase(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
+        _command = new GrantMigratorRoleCommandBaseHandler(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
     }
 
     [Fact]
